Move ZSN tests to test_multivariate.py
